backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
psycopg[binary]
passlib[bcrypt]
python-jose[cryptography]

